requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
dominate==2.6.0
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-Script==2.0.6
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
newsapi==0.1.1
newsapi-python==0.2.6
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
visitor==0.1.3
Werkzeug==1.0.1
